requirements.txt
aiofiles==24.1.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
build==1.0.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Django==4.2.7
dnspython==2.4.2
ecdsa==0.19.0
email-validator==2.1.0.post1
fastapi==0.104.1
Flask==3.0.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector==2.2.9
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.3
passlib==1.7.4
pillow==10.4.0
portalocker==2.8.2
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyodbc==5.0.1
pyperclip==1.9.0
pyproject_hooks==1.0.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
secure-smtplib==0.1.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
starlette==0.27.0
twilio==8.12.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
yarl==1.9.4
